# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: SIAKAD Backend API
  version: 0.1.0
  description: |
    Dokumentasi API awal untuk backend SIAKAD berbasis Go modular monolith.
    Versi ini mencakup endpoint health, discovery modul, serta CRUD `academic_years` dan `semesters`.
servers:
  - url: http://127.0.0.1:18080
    description: Local development
  - url: http://localhost:8080
    description: Default local server
...
paths:
...
  /api/v1/industry-relations/categories:
    get:
      tags: [Industry Relations]
      summary: List industry categories
      parameters:
        - in: query
          name: search
          schema:
            type: string
      responses:
        '200':
          description: Daftar kategori industri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndustryCategoryListResponse'
    post:
      tags: [Industry Relations]
      summary: Buat industry category baru
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IndustryCategoryWriteRequest'
      responses:
        '201':
          description: Kategori industri berhasil dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndustryCategoryDetailResponse'
        '400':
          description: Payload invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Nama kategori bentrok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/industry-relations/categories/{id}:
    get:
      tags: [Industry Relations]
      summary: Detail industry category
      parameters:
        - $ref: '#/components/parameters/IDParam'
      responses:
        '200':
          description: Detail kategori industri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndustryCategoryDetailResponse'
        '400':
          description: ID invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Data tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Industry Relations]
      summary: Update industry category
      parameters:
        - $ref: '#/components/parameters/IDParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IndustryCategoryWriteRequest'
      responses:
        '200':
          description: Kategori industri berhasil diperbarui
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndustryCategoryDetailResponse'
        '400':
          description: Payload invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Data tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Nama kategori bentrok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Industry Relations]
      summary: Hapus industry category
      parameters:
        - $ref: '#/components/parameters/IDParam'
      responses:
        '200':
          description: Kategori industri berhasil dihapus
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: ID invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Data tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/industry-relations/internships:
    get:
      tags: [Industry Relations]
      summary: List internships
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: student_id
          schema:
            type: integer
            format: uint64
        - in: query
          name: company_id
          schema:
            type: integer
            format: uint64
        - in: query
          name: academic_year_id
          schema:
            type: integer
            format: uint64
        - in: query
          name: status
          schema:
            type: string
      responses:
        '200':
          description: Daftar prakerin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternshipListResponse'
        '400':
          description: Parameter invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Industry Relations]
      summary: Buat internship baru
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InternshipWriteRequest'
      responses:
        '201':
          description: Prakerin berhasil dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternshipDetailResponse'
        '400':
          description: Payload invalid atau relasi tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/industry-relations/internships/{id}:
    get:
      tags: [Industry Relations]
      summary: Detail internship
      parameters:
        - $ref: '#/components/parameters/IDParam'
      responses:
        '200':
          description: Detail prakerin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternshipDetailResponse'
        '400':
          description: ID invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
  parameters:
    IDParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: uint64
      description: ID resource
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
      required: [success, error]
...
    MessageResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
...
    IndustryCategory:
      type: object
      properties:
        id:
          type: integer
          format: uint64
        name:
          type: string
          example: Software House
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, name, created_at, updated_at]
    IndustryCategoryWriteRequest:
      type: object
      properties:
        name:
          type: string
          example: Manufaktur
        description:
          type: string
      required: [name]
    IndustryCategoryListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/IndustryCategory'
        meta:
          type: object
          properties:
            total:
              type: integer
          required: [total]
      required: [success, data, meta]
    IndustryCategoryDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/IndustryCategory'
      required: [success, data]
...
    Internship:
      type: object
      properties:
        id:
          type: integer
          format: uint64
        student_id:
          type: integer
          format: uint64
        student_nis:
          type: string
        student_full_name:
          type: string
        company_id:
          type: integer
          format: uint64
        company_name:
          type: string
        academic_year_id:
          type: integer
          format: uint64
        academic_year_name:
          type: string
        start_date:
          type: string
          format: date
          nullable: true
        end_date:
          type: string
          format: date
          nullable: true
        mentor_name:
          type: string
          nullable: true
        status:
          type: string
          example: planned
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - student_id
        - student_nis
        - student_full_name
        - company_id
        - company_name
        - academic_year_id
        - academic_year_name
        - status
        - created_at
        - updated_at
    InternshipWriteRequest:
      type: object
      properties:
        student_id:
          type: integer
          format: uint64
        company_id:
          type: integer
          format: uint64
        academic_year_id:
          type: integer
          format: uint64
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        mentor_name:
          type: string
        status:
          type: string
          example: planned
      required: [student_id, company_id, academic_year_id]
    InternshipListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Internship'
        meta:
          type: object
          properties:
            total:
              type: integer
          required: [total]
      required: [success, data, meta]
    InternshipDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Internship'
      required: [success, data]
...